requirements.txt
streamlit
google-cloud-texttospeech
st-gsheets-connection
pandas
requests
